feat: added profile mechanism
The profile aware beans can be annotated with this mechanism. The ProfileScope is able to get the right profile attached to the bean. ProfileName and the RServer connection properties are now specific for each profile aware collection of beans.

BREAKING: the configuration interface is changed, each deployment needs to update there deployment properties to make this work